requirements.txt
Flask
json5
request3
mysql-connector
gunicorn